Moved DaoEntity, added synchronization to clearCart() method
