+++ src/main/java/com/es/phoneshop/dao/GenericDao.java
@@ -1,5 +1,7 @@
 package com.es.phoneshop.dao;
 
+import com.es.phoneshop.model.DaoEntity;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
