+++ src/main/java/com/es/phoneshop/model/DaoEntity.java
@@ -1,4 +1,4 @@
-package com.es.phoneshop.dao;
+package com.es.phoneshop.model;
 
 public interface DaoEntity<T> {
     Long getId();
