+++ src/main/java/com/es/phoneshop/service/HttpSessionCartService.java
@@ -100,9 +100,13 @@ public class HttpSessionCartService implements CartService {
 
     @Override
     public void clearCart(HttpServletRequest request) {
-        Cart cart = getCart(request);
-        cart.getItems().clear();
-        recalculateCart(cart);
+        HttpSession session = request.getSession();
+
+        synchronized (session) {
+            Cart cart = getCart(request);
+            cart.getItems().clear();
+            recalculateCart(cart);
+        }
     }
 
     private void addExistedInCartProduct(Product product, CartItem cartItem, int quantity) throws OutOfStockException {
